third_party/WebKit/LayoutTests/fast/css/large-integer-value-serializing.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="div"></div>
<script>
test(function(){
    div.style.zIndex = "2147483647";
    assert_equals(div.style.zIndex, "2147483647");
}, "Verify that large CSS integer values do not serialize with scientific notation.");

</script>
